etc/google_images/requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
lxml==3.7.3
numpy==1.12.0
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-pptx==0.6.2
pytz==2016.10
six==1.10.0
XlsxWriter==0.9.6
